issue marcellussiegburg/modelling-tasks#1 Petri net tokens
it works. But there is a big room for improvement:
- I used the binary function `atop` as a parameter for a foldl, because I did not find one that does the same on a list.
- because of the usage of foldl I had to come up with a neutral element. I tried "empty x y" and "nil". Neither of them worked because of their types. Instead, I just did a white outlined circle.

Also, I used "log" for translating the tokens outwards from the center of the circle (the place) because it gives it a more natural feel. Otherwise if I had set a static value, then they would either be too spread apart in case of two tokens, or too close to each other in case of 6 tokens, for example.

I hope the issue referencing is working.