issue marcellussiegburg/modelling-tasks#1 Petro net tokens
Just changed the neutral element that is used in the "foldl" to a very small
white outlined circle.
(circle 0.00001 # lc white)

This is of course not ideal. I can improve it as soon as I learn how to do `atop` on a list of items, similarly to how (+) is binary, but "sum" works on a list, or at least when I learn how to do a transparent color.
Of course, x `atop` y `atop` z is not the same as (x `atop` y) `atop` z in general, but we can treat it as such in our case, because tokens are not stacking or even touching each other.